+++ CHANGELOG.md
@@ -6,9 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Return Siren document containing peer ID, listen addresses and links to `/swaps` and `/swaps/rfc003` on `GET /` with the Accept request HTTP header set to `application/vnd.siren+json`.
+
 ### Changed
 - Write all diagnostics and log messages to stderr.
-- **Breaking (HTTP API):** Return Siren document containing peer ID, listen addresses and links to `/swaps` and `/swaps/rfc003` on `GET /`.
 
 ## [0.5.0] - 2019-12-06